update all tests to use new createMemoryHistory
