+++ modules/__tests__/IndexRedirect-test.js
@@ -1,7 +1,7 @@
 import expect from 'expect'
 import React from 'react'
 import { render, unmountComponentAtNode } from 'react-dom'
-import createHistory from 'history/lib/createMemoryHistory'
+import createHistory from '../createMemoryHistory'
 import IndexRedirect from '../IndexRedirect'
 import Router from '../Router'
 import Route from '../Route'
