+++ modules/__tests__/Link-test.js
@@ -2,7 +2,7 @@ import expect, { spyOn } from 'expect'
 import React, { Component } from 'react'
 import { Simulate } from 'react-addons-test-utils'
 import { render } from 'react-dom'
-import createHistory from 'history/lib/createMemoryHistory'
+import createHistory from '../createMemoryHistory'
 import execSteps from './execSteps'
 import Router from '../Router'
 import Route from '../Route'
